% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_zsc.R
\name{apply_zsc}
\alias{apply_zsc}
\title{Apply Zero-Shot Classification}
\usage{
apply_zsc(model, text, candidate_labels)
}
\arguments{
\item{model}{A pre-trained model that supports zero-shot classification.}

\item{text}{A character vector containing the text samples to classify.}

\item{candidate_labels}{A character vector containing the candidate labels for classification.}
}
\value{
A tibble with columns: \code{text} (input text), scores for each candidate label, and \code{label} (label with the highest score).
}
\description{
This function applies zero-shot classification on a given text using the provided model and candidate labels.
It returns a tidy data frame (tibble) with the input text, scores for each candidate label, and the label
with the highest score for each text.
}
\examples{
\dontrun{
model <- load_zsc_model("MoritzLaurer/mDeBERTa-v3-base-mnli-xnli")
text <- "This is a test."
labels <- c("Label1", "Label2", "Label3")
apply_zsc(model, text, labels)
}
}
