% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_python.R
\name{check_python}
\alias{check_python}
\title{Check Installed Python Version Suitable to run Transformer Library or Not}
\usage{
check_python()
}
\value{
Invisible NULL. The function primarily provides messages and may stop execution if Python is not available.
}
\description{
This function checks the installed version of Python. If Python is not available,
it prompts the user to install version 3.10. It further checks the Python version
to ensure compatibility and provides recommendations.
}
\examples{
\dontrun{
check_python()
}
}
