% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_classification_model.R
\name{load_classification_model}
\alias{load_classification_model}
\title{Load Text Classification Model}
\usage{
load_classification_model(model_path)
}
\arguments{
\item{model_path}{A character string specifying the path to the pre-trained model and tokenizer. It can be a Huggingface Repository.}
}
\value{
A Python object representing the text classification pipeline.
}
\description{
This function load a text classification model using the Python \code{transformers} library.
It takes a model path as an argument and initializes the model and tokenizer from the \code{transformers} library.
}
\examples{
\dontrun{
model <- load_classification_model("path/to/pretrained/model")
}
}
