% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ner_model.R
\name{load_ner_model}
\alias{load_ner_model}
\title{Load Named Entity Recognition (NER) Model}
\usage{
load_ner_model(model_path)
}
\arguments{
\item{model_path}{A character string specifying the path to the pre-trained model and tokenizer. It can be a Huggingface Repository.}
}
\value{
A Python object representing the NER pipeline.
}
\description{
This function loads a NER model using the Python \code{transformers} library.
It takes a model path as an argument and initializes the model and tokenizer from the \code{transformers} library.
}
\examples{
\dontrun{
model <- load_ner_model("https://huggingface.co/dslim/bert-base-NER")
}
}
