% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zsc_model.R
\name{load_zsc_model}
\alias{load_zsc_model}
\title{Load Zero-Shot Classification Model}
\usage{
load_zsc_model(model_path)
}
\arguments{
\item{model_path}{A character string specifying the path to the pre-trained model. It can be a Huggingface Repository.}
}
\value{
A Python object representing the zero-shot classification pipeline.
}
\description{
This function load a zero-shot classification model using the Python \code{transformers} library.
It initializes a zero-shot classification pipeline using a pre-trained model located at the specified \code{model_path}.
}
\examples{
\dontrun{
zsc_model <- load_zsc_model("path/to/pretrained/model")
}
}
