% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{setup_environment}
\alias{setup_environment}
\title{Setup Python Virtual Environment for Using Transformer}
\usage{
setup_environment(env_name, req_txt_path = "requirements.txt")
}
\arguments{
\item{env_name}{Character. Name for the new virtual environment.}

\item{req_txt_path}{Character. Path to the requirements.txt file for package installation.}
}
\value{
Invisible NULL. The function primarily has side effects of setting up and activating the virtual environment.
}
\description{
This function sets up a new Python virtual environment for using Huggingface Transformer library in R.
It checks for the presence of Python, validates the reticulate version,
and installs packages from a provided requirements.txt file.
}
\examples{
\dontrun{
setup_environment("my_env", "requirements.txt")
}
}
