% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_reticulate.R
\name{setup_reticulate}
\alias{setup_reticulate}
\title{Setup Reticulate for Using Transformer}
\usage{
setup_reticulate()
}
\value{
Invisible NULL. The function primarily has side effects of installing and setting up the package.
}
\description{
This function checks and sets up the \code{reticulate} package for using Huggingface Transformer library in R.
It verifies the presence of Python, ensures that \code{reticulate} is installed,
and also checks its version. If not installed or if the version is incorrect,
it installs the specific desired version.
}
\examples{
\dontrun{
setup_reticulate()
}
}
