% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_classification.R
\name{apply_classification}
\alias{apply_classification}
\title{Apply Text Classification}
\usage{
apply_classification(model, text)
}
\arguments{
\item{model}{A pre-trained model that supports text classification.}

\item{text}{A character vector containing the text samples to classify.}
}
\value{
A data frame with the model's outputs.
}
\description{
This function applies text classification on a given text using the provided model.
It returns a data frame with the model's outputs.
}
\examples{
\dontrun{
model <- load_ner_model("lxyuan/distilbert-base-multilingual-cased-sentiments-student")
text <- c("This is a test.", "Another sample text.")
apply_classification(model, text)
}
}
